add changes

+++ libft.h
@@ -47,7 +47,7 @@ int		ft_memcmp(const void *s1, const void *s2, size_t n);
 int		ft_atoi(const char *str);
 void	ft_putnbr_fd(int n, int fd);
 void	*ft_calloc(int nmemb, int size);
-char	*ft_substr(char const *s, unsigned int start, int len);
+char	*ft_substr(char const *s, unsigned int start, size_t len);
 char	*ft_strjoin(char const *s1, char const *s2);
 char	*ft_strtrim(char const *s1, char const *set);
 char	**ft_split(char const *s, char c);
